+++ src/libs/routes.js
@@ -24,9 +24,9 @@ import * as Upload from 'src/pages/Upload'
 import * as WorkflowsList from 'src/pages/workflows/List'
 import * as WorkflowDetails from 'src/pages/workflows/workflow/WorkflowDetails'
 import * as WorkspaceList from 'src/pages/workspaces/List'
-import * as Analysis from 'src/pages/workspaces/workspace/Analyses'
+import * as Analysis from 'src/pages/workspaces/workspace/analysis/Analyses'
 import * as AnalysisLauncher from 'src/pages/workspaces/workspace/analysis/AnalysisLauncher'
-import * as AppLauncher from 'src/pages/workspaces/workspace/applications/AppLauncher'
+import * as AppLauncher from 'src/pages/workspaces/workspace/analysis/AppLauncher'
 import * as Dashboard from 'src/pages/workspaces/workspace/Dashboard'
 import * as Data from 'src/pages/workspaces/workspace/Data'
 import * as JobHistory from 'src/pages/workspaces/workspace/JobHistory'
