+++ src/pages/workspaces/workspace/analysis/Analyses.js
@@ -3,7 +3,6 @@ import _ from 'lodash/fp'
 import * as qs from 'qs'
 import { Fragment, useEffect, useState } from 'react'
 import { a, div, h, img, span } from 'react-hyperscript-helpers'
-import { AnalysisModal } from 'src/components/AnalysisModal'
 import * as breadcrumbs from 'src/components/breadcrumbs'
 import { requesterPaysWrapper, withRequesterPaysHandler } from 'src/components/bucket-utils'
 import { withViewToggle } from 'src/components/CardsListToggle'
@@ -34,7 +33,8 @@ import { authStore } from 'src/libs/state'
 import * as StateHistory from 'src/libs/state-history'
 import * as Style from 'src/libs/style'
 import * as Utils from 'src/libs/utils'
-import { AzureComputeModal } from 'src/pages/workspaces/workspace/analysis/AzureComputeModal'
+import { AnalysisModal } from 'src/pages/workspaces/workspace/analysis/modals/AnalysisModal'
+import { AzureComputeModal } from 'src/pages/workspaces/workspace/analysis/modals/AzureComputeModal'
 import ExportAnalysisModal from 'src/pages/workspaces/workspace/analysis/modals/ExportAnalysisModal'
 import { wrapWorkspace } from 'src/pages/workspaces/workspace/WorkspaceContainer'
