+++ src/pages/workspaces/workspace/analysis/ContextBar.js
@@ -7,6 +7,7 @@ import Interactive from 'src/components/Interactive'
 import { getAppType, tools } from 'src/components/notebook-utils'
 import { getRegionInfo } from 'src/components/region-common'
 import { appLauncherTabName } from 'src/components/runtime-common'
+import TooltipTrigger from 'src/components/TooltipTrigger'
 import cloudIcon from 'src/icons/cloud-compute.svg'
 import cromwellImg from 'src/images/cromwell-logo.png' // To be replaced by something square
 import galaxyLogo from 'src/images/galaxy-project-logo-square.png'
@@ -20,8 +21,6 @@ import { getCostDisplayForDisk, getCostDisplayForTool, getCurrentApp, getCurrent
 import * as Style from 'src/libs/style'
 import * as Utils from 'src/libs/utils'
 import { CloudEnvironmentModal } from 'src/pages/workspaces/workspace/analysis/modals/CloudEnvironmentModal'
-
-import TooltipTrigger from './TooltipTrigger'
 
 
 const contextBarStyles = {
