+++ src/pages/workspaces/workspace/analysis/modals/AnalysisModal.js
@@ -2,7 +2,6 @@ import _ from 'lodash/fp'
 import { Fragment, useState } from 'react'
 import { div, h, h2, hr, img, span } from 'react-hyperscript-helpers'
 import { ButtonPrimary, Clickable, IdContainer, Select, WarningTitle } from 'src/components/common'
-import { ComputeModalBase } from 'src/components/ComputeModal'
 import { CromwellModalBase } from 'src/components/CromwellModal'
 import Dropzone from 'src/components/Dropzone'
 import { GalaxyModalBase } from 'src/components/GalaxyModal'
@@ -26,6 +25,7 @@ import { usePrevious, withDisplayName } from 'src/libs/react-utils'
 import { getCurrentApp, getCurrentRuntime, isResourceDeletable } from 'src/libs/runtime-utils'
 import * as Style from 'src/libs/style'
 import * as Utils from 'src/libs/utils'
+import { ComputeModalBase } from 'src/pages/workspaces/workspace/analysis/modals/ComputeModal'
 import validate from 'validate.js'
