+++ src/pages/workspaces/workspace/analysis/modals/CloudEnvironmentModal.js
@@ -2,14 +2,12 @@ import _ from 'lodash/fp'
 import { Fragment, useState } from 'react'
 import { div, h, hr, img, span } from 'react-hyperscript-helpers'
 import { Clickable, spinnerOverlay } from 'src/components/common'
-import { ComputeModalBase } from 'src/components/ComputeModal'
 import { CromwellModalBase } from 'src/components/CromwellModal'
 import { GalaxyModalBase } from 'src/components/GalaxyModal'
 import { icon } from 'src/components/icons'
 import ModalDrawer from 'src/components/ModalDrawer'
 import { getAppType, getToolsToDisplay, isPauseSupported, isToolAnApp, tools } from 'src/components/notebook-utils'
 import { appLauncherTabName } from 'src/components/runtime-common'
-import { AppErrorModal, RuntimeErrorModal } from 'src/components/RuntimeManager'
 import TitleBar from 'src/components/TitleBar'
 import cromwellImg from 'src/images/cromwell-logo.png'
 import galaxyLogo from 'src/images/galaxy-logo.svg'
@@ -28,7 +26,9 @@ import {
 } from 'src/libs/runtime-utils'
 import { cookieReadyStore } from 'src/libs/state'
 import * as Utils from 'src/libs/utils'
-import { AzureComputeModalBase } from 'src/pages/workspaces/workspace/analysis/AzureComputeModal'
+import { AzureComputeModalBase } from 'src/pages/workspaces/workspace/analysis/modals/AzureComputeModal'
+import { ComputeModalBase } from 'src/pages/workspaces/workspace/analysis/modals/ComputeModal'
+import { AppErrorModal, RuntimeErrorModal } from 'src/pages/workspaces/workspace/analysis/RuntimeManager'
 
 
 const titleId = 'cloud-env-modal'
